requirements.txt
pydantic-ai>=0.0.14
httpx>=0.27.0
pydantic>=2.0.0
torch>=2.0.0
transformers>=4.40.0
pillow>=10.0.0
accelerate>=0.30.0
numpy==1.26.0
vllm==0.8.5
